# crates/mantissa-rest/openapi/mantissa-rest.openapi.json
{
  "components": {
    "schemas": {
...
      "AgentInputRequest": {
        "additionalProperties": false,
        "description": "REST request body for queuing input on an agent session.",
        "properties": {
          "input": {
            "type": "string"
          }
        },
        "required": [
          "input"
        ],
        "type": "object"
      },
      "AgentInputResponse": {
        "description": "REST response returned after queuing input on an agent session.",
        "properties": {
          "accepted": {
            "type": "boolean"
          }
        },
        "required": [
          "accepted"
        ],
        "type": "object"
      },
...
      "AgentRunSummary": {
        "description": "REST-facing compact agent run row.",
        "properties": {
          "execution_platform": {
            "type": "string"
          },
          "exit_code": {
            "format": "int32",
            "type": [
              "integer",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "isolation_mode": {
            "type": "string"
          },
          "isolation_profile": {
            "type": [
              "string",
              "null"
            ]
          },
          "session_name": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          },
          "workload_id": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "id",
          "session_name",
          "status",
          "execution_platform",
          "isolation_mode",
          "updated_at"
        ],
        "type": "object"
      },
...
      "AgentSessionDetail": {
        "description": "REST-facing detailed agent inspection response.",
        "properties": {
          "runs": {
            "items": {
              "$ref": "#/components/schemas/AgentRun"
            },
            "type": "array"
          },
          "session": {
            "$ref": "#/components/schemas/AgentSession"
          }
        },
        "required": [
          "session",
          "runs"
        ],
        "type": "object"
      },
      "AgentSessionSummary": {
        "description": "REST-facing compact agent session row.",
        "properties": {
          "active_run_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "execution_platform": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "isolation_mode": {
            "type": "string"
          },
          "isolation_profile": {
            "type": [
              "string",
              "null"
            ]
          },
          "last_run_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "status",
          "execution_platform",
          "isolation_mode",
          "updated_at"
        ],
        "type": "object"
      },
      "AgentSubmitRequest": {
        "additionalProperties": false,
        "description": "REST request body for submitting one durable agent session.",
        "properties": {
          "manifest": {
            "$ref": "#/components/schemas/AgentManifest"
          }
        },
        "required": [
          "manifest"
        ],
        "type": "object"
      },
      "AgentSubmitResponse": {
        "description": "REST response returned after submitting one durable agent session.",
        "properties": {
          "cpu_millis": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "execution_platform": {
            "type": "string"
          },
          "gpu_count": {
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "image": {
            "type": "string"
          },
          "isolation_mode": {
            "type": "string"
          },
          "isolation_profile": {
            "type": [
              "string",
              "null"
            ]
          },
          "memory_mib": {
            "format": "int64",
            "minimum": 0,
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "session_id": {
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "name",
          "image",
          "cpu_millis",
          "memory_mib",
          "gpu_count",
          "execution_platform",
          "isolation_mode"
        ],
        "type": "object"
      },
...
      "RestErrorBody": {
        "description": "Stable JSON error body returned by REST handlers.",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "Mantissa REST token",
        "description": "Daemon-generated REST token presented as a Bearer token.",
        "scheme": "bearer",
        "type": "http"
      },
      "mtls": {
        "description": "Required when the REST listener is bound to a non-loopback interface.",
        "type": "mutualTLS"
      }
    }
  },
  "info": {
    "description": "Local HTTP facade over the Mantissa Cap'n Proto admin session. All /v1 endpoints require an Authorization: Bearer token. Non-loopback listeners must also be served with TLS and client certificate verification enabled.",
    "license": {
      "identifier": "Apache-2.0",
      "name": "Apache-2.0"
    },
    "title": "Mantissa REST API",
    "version": "0.1.0"
  },
  "openapi": "3.2.0",
  "paths": {
...
    "/v1/agents/sessions": {
      "get": {
        "operationId": "list_sessions",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AgentSessionSummary"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Durable agent sessions."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Invalid request syntax, query parameters, or resource selector."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Missing or invalid REST bearer token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Requested resource was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Request conflicts with the current resource state."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Unexpected REST facade or local client failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Local daemon or REST worker is unavailable."
          }
        },
        "summary": "Lists durable agent sessions visible to the local daemon.",
        "tags": [
          "agents"
        ]
      },
      "post": {
        "operationId": "submit_session",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentSubmitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentSubmitResponse"
                }
              }
            },
            "description": "Submitted agent session metadata."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Invalid request syntax, query parameters, or resource selector."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Missing or invalid REST bearer token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Requested resource was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Request conflicts with the current resource state."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Unexpected REST facade or local client failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Local daemon or REST worker is unavailable."
          }
        },
        "summary": "Submits one durable agent session manifest to the local daemon.",
        "tags": [
          "agents"
        ]
      }
    },
    "/v1/agents/sessions/{session_id}": {
      "delete": {
        "operationId": "delete_session",
        "parameters": [
          {
            "description": "Agent session UUID string or exact session name.",
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentSession"
                }
              }
            },
            "description": "Deleted agent session snapshot."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Invalid request syntax, query parameters, or resource selector."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Missing or invalid REST bearer token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Requested resource was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Request conflicts with the current resource state."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Unexpected REST facade or local client failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Local daemon or REST worker is unavailable."
          }
        },
        "summary": "Deletes one closed durable agent session and its retained run history.",
        "tags": [
          "agents"
        ]
      },
      "get": {
        "operationId": "get_session",
        "parameters": [
          {
            "description": "Agent session UUID string or exact session name.",
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentSessionDetail"
                }
              }
            },
            "description": "Agent session detail."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Invalid request syntax, query parameters, or resource selector."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Missing or invalid REST bearer token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Requested resource was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Request conflicts with the current resource state."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Unexpected REST facade or local client failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Local daemon or REST worker is unavailable."
          }
        },
        "summary": "Fetches one durable agent session and its retained run history.",
        "tags": [
          "agents"
        ]
      }
    },
    "/v1/agents/sessions/{session_id}/cancel": {
      "post": {
        "operationId": "cancel_session",
        "parameters": [
          {
            "description": "Agent session UUID string or exact session name.",
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentSession"
                }
              }
            },
            "description": "Updated agent session snapshot."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Invalid request syntax, query parameters, or resource selector."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Missing or invalid REST bearer token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Requested resource was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Request conflicts with the current resource state."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Unexpected REST facade or local client failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Local daemon or REST worker is unavailable."
          }
        },
        "summary": "Requests cancellation for one active or queued agent session run.",
        "tags": [
          "agents"
        ]
      }
    },
    "/v1/agents/sessions/{session_id}/close": {
      "post": {
        "operationId": "close_session",
        "parameters": [
          {
            "description": "Agent session UUID string or exact session name.",
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentSession"
                }
              }
            },
            "description": "Updated agent session snapshot."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Invalid request syntax, query parameters, or resource selector."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Missing or invalid REST bearer token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Requested resource was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Request conflicts with the current resource state."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Unexpected REST facade or local client failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Local daemon or REST worker is unavailable."
          }
        },
        "summary": "Closes one durable agent session and rejects future input.",
        "tags": [
          "agents"
        ]
      }
    },
    "/v1/agents/sessions/{session_id}/input": {
      "post": {
        "operationId": "submit_input",
        "parameters": [
          {
            "description": "Agent session UUID string or exact session name.",
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentInputRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentInputResponse"
                }
              }
            },
            "description": "Input was accepted for the session."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Invalid request syntax, query parameters, or resource selector."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Missing or invalid REST bearer token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Requested resource was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Request conflicts with the current resource state."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Unexpected REST facade or local client failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Local daemon or REST worker is unavailable."
          }
        },
        "summary": "Queues structured input on one idle agent session.",
        "tags": [
          "agents"
        ]
      }
    },
    "/v1/agents/sessions/{session_id}/runs": {
      "get": {
        "operationId": "list_runs",
        "parameters": [
          {
            "description": "Agent session UUID string or exact session name.",
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AgentRunSummary"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Agent runs owned by the session."
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Invalid request syntax, query parameters, or resource selector."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Missing or invalid REST bearer token."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Requested resource was not found."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Request conflicts with the current resource state."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Unexpected REST facade or local client failure."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RestErrorBody"
                }
              }
            },
            "description": "Local daemon or REST worker is unavailable."
          }
        },
        "summary": "Lists durable runs for one agent session.",
        "tags": [
          "agents"
        ]
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
}